+++ packages/applets/calculator/src/_types/IEvaluationResult.ts
@@ -1,0 +1,16 @@
+import {ISyntaxError} from "./errors/ISyntaxError";
+
+export type IEvaluationResult =
+    | {
+          error: ISyntaxError | Error;
+      }
+    | {
+          /** The evaluated expression, which may include fixes */
+          expression: string;
+          /** The result */
+          result: {
+              text: string;
+              raw: any;
+              formatted: JSX.Element;
+          };
+      };