+++ CHANGELOG.md
@@ -14,3 +14,7 @@
 -   Fixed syntax highlighting updating [#159](https://github.com/LaunchMenu/LaunchMenu/pull/159)
 -   Added standard item highlighting customization [#160](https://github.com/LaunchMenu/LaunchMenu/pull/160)
 -   Added react error boundaries [#164](https://github.com/LaunchMenu/LaunchMenu/pull/164)
+
+#### applet-calculator
+
+-   Created initial simple calculator implementation [#163](https://github.com/LaunchMenu/LaunchMenu/pull/163)